refactor: create /format folder
