+++ src/format/formatFailureMessageTraceLine.js
@@ -1,0 +1,6 @@
+const chalk = require('chalk');
+
+const formatFailureMessageTraceLine = (description, relativeFilePath, row, column) =>
+  chalk`${description}({cyan ${relativeFilePath}}:{black.bold ${row}}:{black.bold ${column}})`;
+
+module.exports = formatFailureMessageTraceLine;
